Convert to long long

+++ BigInt.h
@@ -25,7 +25,7 @@ class BigInt {
 	friend BigInt operator+(const BigInt& a,const BigInt& b);
 	friend BigInt operator-(const BigInt& a, const BigInt& b);
 	friend BigInt subtract(BigInt& a, BigInt& b,bool& isNegative);
-	friend BigInt add(std::vector<int>& a, std::vector<int>& b,bool& isNegative);
+	friend BigInt add(std::vector<long long>& a, std::vector<long long>& b,bool& isNegative);
 	friend BigInt operator/(const BigInt& a, const BigInt& b);
 	friend BigInt operator*(const BigInt& a, const BigInt& b);
 	
@@ -84,12 +84,12 @@ class BigInt {
 
 	private:
 		void eraseLeadingZeros();
-	int size() const{
+	long long size() const{
 		return digits.size();
 	}
 	
 	bool isNegative;
-	std::vector<int> digits;
+	std::vector<long long> digits;
 
 };
